+++ DotFeather/Time.cs
@@ -2,21 +2,21 @@ using System;
 namespace DotFeather
 {
 	/// <summary>
-	/// 時間に関係する情報を提供します。
+	/// Provides time related information.
 	/// </summary>
 	public static class Time
 	{
 		/// <summary>
-		/// ゲームが起動してからの時刻を取得します。
+		/// Get the time since the game started.
 		/// </summary>
-		public static double Now { get; internal set; }
+		public static float Now { get; internal set; }
 		/// <summary>
-		/// 前回フレームとの差分時間を取得します。
+		/// Get the delta time from the previous frame.
 		/// </summary>
-		public static double DeltaTime { get; internal set; }
+		public static float DeltaTime { get; internal set; }
 
 		/// <summary>
-		/// 現在のフレームレートを取得します。
+		/// Get the current frame rate.
 		/// </summary>
 		public static int Fps { get; internal set; }
 	}
